refactor(tool): move agent tool schemas out of skill module
