+++ packages/global/core/ai/skill/type.ts
@@ -1,12 +1,9 @@
 import z from 'zod';
-import { NodeToolConfigTypeSchema } from '../../workflow/type/node';
-import { FlowNodeInputTypeEnum } from '../../workflow/node/constant';
 import {
   AgentSkillSourceEnum,
   AgentSkillCategoryEnum,
   AgentSkillTypeEnum,
-  AgentSkillCreationStatusEnum,
-  AgentToolInputModeEnum
+  AgentSkillCreationStatusEnum
 } from './constants';
 import { SandboxStatusEnum, SandboxTypeEnum } from '../sandbox/constants';
 import { SandboxImageConfigSchema } from '../sandbox/type';
@@ -175,45 +172,3 @@ export const SandboxInstanceSchema = z.object({
   detail: SandboxInstanceDetailSchema
 });
 export type SandboxInstanceSchemaType = z.infer<typeof SandboxInstanceSchema>;
-
-const AgentToolInputConfigValueSchema = z.object({
-  key: z.string(),
-  mode: z.enum(AgentToolInputModeEnum)
-});
-
-/**
- * Agent 工具只持久化参数输入来源。预处理兼容 selectedType 上线期间产生的临时快照，
- * 解析结果始终收敛为 key + mode，避免 runtime 继续依赖工作流渲染协议。
- */
-export const AgentToolInputConfigSchema = z.preprocess((value) => {
-  if (!value || typeof value !== 'object') return value;
-
-  const input = value as Record<string, unknown>;
-  if (input.mode !== undefined) return value;
-  if (typeof input.key !== 'string') return value;
-
-  const renderTypeList = Array.isArray(input.renderTypeList) ? input.renderTypeList : [];
-  const selectedType =
-    input.selectedType ??
-    (typeof input.selectedTypeIndex === 'number'
-      ? renderTypeList[input.selectedTypeIndex]
-      : undefined);
-
-  return {
-    key: input.key,
-    mode:
-      selectedType === FlowNodeInputTypeEnum.agentGenerated
-        ? AgentToolInputModeEnum.agentGenerated
-        : AgentToolInputModeEnum.manual
-  };
-}, AgentToolInputConfigValueSchema);
-export type AgentToolInputConfigType = z.infer<typeof AgentToolInputConfigSchema>;
-
-export const SkillToolSchema = z.object({
-  id: z.string(),
-  source: z.string().optional(),
-  toolConfig: NodeToolConfigTypeSchema.optional(),
-  inputs: z.array(AgentToolInputConfigSchema).optional(),
-  config: z.record(z.string(), z.any())
-});
-export type SkillToolType = z.infer<typeof SkillToolSchema>;
