+++ packages/global/core/app/tool/constants.ts
@@ -11,3 +11,8 @@ export enum AppToolSourceEnum {
    */
   community = 'community'
 }
+
+export enum AgentToolInputModeEnum {
+  agentGenerated = 'agentGenerated',
+  manual = 'manual'
+}
